+++ lib/moxl/src/Stanza/Bookmark2.php
@@ -6,26 +6,26 @@ use App\Conference;
 
 class Bookmark2
 {
-    public static function get()
+    public static function get($version = '1')
     {
         $dom = new \DOMDocument('1.0', 'UTF-8');
         $pubsub = $dom->createElementNS('http://jabber.org/protocol/pubsub', 'pubsub');
 
         $items = $dom->createElement('items');
-        $items->setAttribute('node', 'urn:xmpp:bookmarks:0');
+        $items->setAttribute('node', 'urn:xmpp:bookmarks:'.$version);
         $pubsub->appendChild($items);
 
         $xml = \Moxl\API::iqWrapper($pubsub, false, 'get');
         \Moxl\API::request($xml);
     }
 
-    public static function set(Conference $conf)
+    public static function set(Conference $conf, $version = '1')
     {
         $dom = new \DOMDocument('1.0', 'UTF-8');
         $pubsub = $dom->createElementNS('http://jabber.org/protocol/pubsub', 'pubsub');
 
         $publish = $dom->createElement('publish');
-        $publish->setAttribute('node', 'urn:xmpp:bookmarks:0');
+        $publish->setAttribute('node', 'urn:xmpp:bookmarks:'.$version);
         $pubsub->appendChild($publish);
 
         $item = $dom->createElement('item');
@@ -33,7 +33,7 @@ class Bookmark2
         $publish->appendChild($item);
 
         $conference = $dom->createElement('conference');
-        $conference->setAttribute('xmlns', 'urn:xmpp:bookmarks:0');
+        $conference->setAttribute('xmlns', 'urn:xmpp:bookmarks:'.$version);
         $conference->setAttribute('name', $conf->name);
         if ($conf->autojoin) {
             $conference->setAttribute('autojoin', 'true');
@@ -42,6 +42,24 @@ class Bookmark2
 
         $nick = $dom->createElement('nick', $conf->nick);
         $conference->appendChild($nick);
+
+        if ($conf->extensions) {
+            $domExtensions = new \DOMDocument('1.0', 'UTF-8');
+            $domExtensions->loadXML($conf->extensions);
+
+            $extensions = $dom->importNode($domExtensions->documentElement, true);
+            $conference->appendChild($extensions);
+        } else if ($conf->notify !== null) {
+            $extensions = $dom->createElement('extensions');
+            $conference->appendChild($extensions);
+        }
+
+        if ($conf->notify !== null) {
+            $notifications = $dom->createElement('notifications');
+            $notifications->setAttribute('xmlns', Conference::$xmlns);
+            $notifications->setAttribute('notify', $conf->notificationKey);
+            $extensions->appendChild($notifications);
+        }
 
         // Publish option
         $publishOption = $dom->createElement('publish-option');