+++ lib/moxl/src/Xec/Action/Bookmark2/Set.php
@@ -9,11 +9,21 @@ use Moxl\Stanza\Bookmark2;
 class Set extends Action
 {
     protected $_conference;
+    protected $_version = '1';
 
     public function request()
     {
         $this->store();
-        Bookmark2::set($this->_conference);
+
+        // If we set the version 1, ensure to remove 0
+        if ($this->_version == '1' && $this->_conference->bookmarkversion == 0) {
+            $d = new Delete;
+            $d->setId($this->_conference->conference)
+              ->setVersion('0')
+              ->request();
+        }
+
+        Bookmark2::set($this->_conference, $this->_version);
     }
 
     public function setConference(Conference $conference)