+++ CHANGELOG.md
@@ -36,6 +36,7 @@ v0.18 (trunk)
  * Improve the Draw widget lines quality (christine-ho-dev)
  * New emoji picker (christine-ho-dev)
  * Add on-the-fly picture compression for the Picture proxy for larger pictures
+ * Update XEP-0402 to urn:xmpp:bookmarks:1 and add xmpp:movim.eu/notifications:0 extension support
 
 v0.17.1
 ---------------------------